+++ src/summarise.cpp
@@ -44,6 +44,7 @@ SEXP reconstruct_groups(const DataFrame& old_groups, const List& new_indices, co
 
   set_rownames(out, new_indices.size());
   set_class(out, NaturalDataFrame::classes());
+  copy_attrib(out, old_groups, symbols::dot_drop);
   out.attr("names") = names;
   return out ;
 }