+++ tests/testthat/test-count-tally.r
@@ -184,3 +184,44 @@ test_that("returns error if user-defined name equals a grouped variable", {
   expect_error(df %>% group_by(val) %>% add_tally(name = name))
 })
 
+# count and .drop ----------------------------------------------------
+
+test_that("count() deals with .drop", {
+  d <- tibble(
+    f1 = factor("b", levels = c("a", "b", "c")),
+    f2 = factor("g", levels = c("e", "f", "g")),
+    x  = 48
+  )
+  res <- d %>%
+    group_by(f1, .drop = TRUE) %>%
+    count(f2, .drop = TRUE)
+
+  res2 <- d %>%
+    group_by(f1, .drop = TRUE) %>%
+    count(f2)
+
+  res3 <- d %>%
+    group_by(f1, .drop = TRUE) %>%
+    count(f2, .drop = FALSE)
+
+  expect_equal(n_groups(res), 1L)
+  expect_identical(res, res2)
+  expect_equal(n_groups(res3), 3L)
+  expect_equal(nrow(res3), 9L)
+})
+
+test_that("add_count() respects .drop",  {
+  d <- tibble(
+    f1 = factor("b", levels = c("a", "b", "c")),
+    f2 = factor("g", levels = c("e", "f", "g")),
+    x  = 48
+  )
+  res1 <- d %>% group_by(f1) %>% add_count(f2, .drop = FALSE)
+  res2 <- d %>% group_by(f1) %>% add_count(f2, .drop = TRUE)
+  res3 <- d %>% group_by(f1) %>% add_count(f2)
+
+  expect_identical(res2, res3)
+  expect_equal(n_groups(res2), 1)
+  expect_equal(n_groups(res1), 3)
+})
+