+++ tests/testthat/test-empty-groups.R
@@ -1,11 +1,11 @@
 context("empty groups")
 
-df <- tibble(
+df <- data_frame(
   e = 1,
   f = factor(c(1, 1, 2, 2), levels = 1:3),
   g = c(1, 1, 2, 2),
   x = c(1, 2, 1, 4)) %>%
-  group_by(e, f, g)
+  group_by(e, f, g, .drop = FALSE)
 
 test_that("filter and slice keep zero length groups", {
   expect_equal(group_size(filter(df, f == 1)), c(2, 0, 0) )
@@ -61,6 +61,18 @@ test_that("joins respect zero length groups", {
   df2 <- tibble(f = factor( c(2,2,3,3), levels = 1:3), y = c(1,2,3,4)) %>%
     group_by(f)
 
+  expect_equal(group_size(left_join( df1, df2, by = "f")),  c(2,4))
+  expect_equal(group_size(right_join( df1, df2, by = "f")),  c(4,2))
+  expect_equal(group_size(full_join( df1, df2, by = "f")),  c(2,4,2))
+  expect_equal(group_size(anti_join( df1, df2, by = "f")),  c(2))
+  expect_equal(group_size(inner_join( df1, df2, by = "f")),  c(4))
+
+
+  df1 <- tibble(f = factor( c(1,1,2,2), levels = 1:3), x = c(1,2,1,4)) %>%
+    group_by(f, .drop = FALSE)
+  df2 <- tibble(f = factor( c(2,2,3,3), levels = 1:3), y = c(1,2,3,4)) %>%
+    group_by(f, .drop = FALSE)
+
   expect_equal(group_size(left_join( df1, df2, by = "f")),  c(2,4,0))
   expect_equal(group_size(right_join( df1, df2, by = "f")),  c(0,4,2))
   expect_equal(group_size(full_join( df1, df2, by = "f")),  c(2,4,2))
@@ -69,7 +81,7 @@ test_that("joins respect zero length groups", {
 })
 
 test_that("n_groups respects zero-length groups (#341)", {
-  df <- tibble(x = factor(1:3, levels = 1:4)) %>% group_by(x)
+  df <- tibble(x = factor(1:3, levels = 1:4)) %>% group_by(x, .drop = FALSE)
   expect_equal(n_groups(df), 4)
 })
 
@@ -77,7 +89,7 @@ test_that("summarise respects zero-length groups (#341)", {
   df <- tibble(x = factor(rep(1:3, each = 10), levels = 1:4))
 
   out <- df %>%
-    group_by(x) %>%
+    group_by(x, .drop = FALSE) %>%
     summarise(n = n())
 
   expect_equal(out$n, c(10L, 10L, 10L, 0L))