+++ tests/testthat/test-group-by.r
@@ -89,6 +89,9 @@ test_that("group_by handles NA in factors #341", {
     x  = 1:4
   )
   expect_warning(g <- group_by(d, f1, f2))
+  expect_equal(group_size(g), c(1L,1L,1L,1L))
+
+  expect_warning(g <- group_by(d, f1, f2, .drop = FALSE))
   expect_equal(group_size(g), c(1L,1L,1L,0L,1L))
 })
 
@@ -117,9 +120,19 @@ test_that("group_by only applies the allow list to grouping variables", {
   df <- data.frame(times = 1:5, x = 1:5)
   df$times <- as.POSIXlt(seq.Date(Sys.Date(), length.out = 5, by = "day"))
 
+  res <- group_by(df, x, .drop = FALSE)
+  expect_equal(groups(res), list(sym("x")))
+  expect_identical(
+    group_data(res),
+    structure(tibble(x := 1:5, ".rows" := as.list(1:5)), .drop = FALSE)
+  )
+
   res <- group_by(df, x)
   expect_equal(groups(res), list(sym("x")))
-  expect_identical(group_data(res), tibble(x := 1:5, ".rows" := as.list(1:5)))
+  expect_identical(
+    group_data(res),
+    structure(tibble(x := 1:5, ".rows" := as.list(1:5)), .drop = TRUE)
+  )
 })
 
 test_that("group_by fails when lists are used as grouping variables (#276)", {
@@ -150,7 +163,7 @@ test_that("grouped_df errors on NULL labels (#398)", {
 test_that("grouped_df errors on non-existent var (#2330)", {
   df <- data.frame(x = 1:5)
   expect_error(
-    grouped_df(df, list(quote(y))),
+    grouped_df(df, list(quote(y)), FALSE),
     "Column `y` is unknown"
   )
 })
@@ -165,14 +178,15 @@ test_that("group_by only creates one group for NA (#401)", {
 })
 
 test_that("there can be 0 groups (#486)", {
-  data <- data.frame(a = numeric(0), g = character(0)) %>% group_by(g)
+  data <- tibble(a = numeric(0), g = character(0)) %>% group_by(g)
   expect_equal(length(data$a), 0L)
   expect_equal(length(data$g), 0L)
   expect_equal(map_int(group_rows(data), length), integer(0))
 })
 
 test_that("group_by works with zero-row data frames (#486)", {
-  dfg <- group_by(data.frame(a = numeric(0), b = numeric(0), g = character(0)), g)
+  df <- data.frame(a = numeric(0), b = numeric(0), g = character(0))
+  dfg <- group_by(df, g, .drop = FALSE)
   expect_equal(dim(dfg), c(0, 3))
   expect_groups(dfg, "g")
   expect_equal(group_size(dfg), integer(0))
@@ -391,4 +405,91 @@ test_that("group_by() with empty spec produces a grouped data frame with 0 group
   gdata <- group_data(group_by(iris))
   expect_equal(names(gdata), ".rows")
   expect_equal(gdata$.rows, list(1:nrow(iris)))
+})
+
+# .drop = TRUE ---------------------------------------------------
+
+test_that("group_by(.drop = TRUE) drops empty groups (4061)", {
+  res <- iris %>%
+    filter(Species == "setosa") %>%
+    group_by(Species, .drop = TRUE)
+
+  expect_identical(
+    group_data(res),
+    structure(
+      tibble(Species = factor("setosa", levels = levels(iris$Species)), .rows := list(1:50)),
+      .drop = TRUE
+    )
+  )
+
+  expect_true(group_drops(res))
+})
+
+test_that("grouped data frames remember their .drop (#4061)", {
+  res <- iris %>%
+    filter(Species == "setosa") %>%
+    group_by(Species, .drop = TRUE)
+
+  res2 <- res %>%
+    filter(Sepal.Length > 5)
+  expect_true(group_drops(res2))
+
+  res3 <- res %>%
+    filter(Sepal.Length > 5, .preserve = FALSE)
+  expect_true(group_drops(res3))
+
+  res4 <- res3 %>%
+    group_by(Species)
+  expect_true(group_drops(res4))
+  expect_equal(nrow(group_data(res4)), 1L)
+})
+
+test_that("summarise maintains the .drop attribute (#4061)", {
+  df <- tibble(
+    f1 = factor("a", levels = c("a", "b", "c")),
+    f2 = factor("d", levels = c("d", "e", "f", "g")),
+    x  = 42
+  )
+
+  res <- df %>%
+    group_by(f1, f2, .drop = TRUE)
+  expect_equal(n_groups(res), 1L)
+
+  res2 <- summarise(res, x = sum(x))
+  expect_equal(n_groups(res2), 1L)
+  expect_true(group_drops(res2))
+})
+
+test_that("joins maintain the .drop attribute (#4061)", {
+  df1 <- group_by(tibble(
+    f1 = factor(c("a", "b"), levels = c("a", "b", "c")),
+    x  = 42:43
+  ), f1, .drop = TRUE)
+
+  df2 <- group_by(tibble(
+    f1 = factor(c("a"), levels = c("a", "b", "c")),
+    y = 1
+  ), f1, .drop = TRUE)
+
+  res <- left_join(df1, df2)
+  expect_equal(n_groups(res), 2L)
+
+  df2 <- group_by(tibble(
+    f1 = factor(c("a", "c"), levels = c("a", "b", "c")),
+    y = 1:2
+  ), f1, .drop = TRUE)
+  res <- full_join(df1, df2)
+  expect_equal(n_groups(res), 3L)
+})
+
+test_that("group_by(add = TRUE) sets .drop if the origonal data was .drop", {
+  d <- tibble(
+    f1 = factor("b", levels = c("a", "b", "c")),
+    f2 = factor("g", levels = c("e", "f", "g")),
+    x  = 48
+  )
+
+  res <- group_by(group_by(d, f1, .drop = TRUE), f2, add = TRUE)
+  expect_equal(n_groups(res), 1L)
+  expect_true(group_drops(res))
 })